+++ lib/cal.js
@@ -1,3 +1,5 @@
+const utils = require('./utils')
+
 function calculateInteractiveNodes(properties) {
   return properties.reduce(
     (acc, p) => acc.concat(calculateInteractiveNodesPerProperty(p)),
@@ -15,9 +17,6 @@ function calculateInteractiveNodesPerProperty(p) {
     1,
     p
   )
-  newContracts.forEach(r => {
-    genIsValidChallenge(r.statement, name, 1)
-  })
   return newContracts
 }
 
@@ -28,7 +27,7 @@ function calculateInteractiveNodesPerProperty(p) {
  * @param {*} index
  */
 function searchInteractiveNode(contracts, property, name, index, parent) {
-  if (isNotAtomicProposition(property.predicate)) {
+  if (utils.isNotAtomicProposition(property.predicate)) {
     const newContract = {
       dec: {
         predicate: name + '_' + index + '_compile',
@@ -42,6 +41,15 @@ function searchInteractiveNode(contracts, property, name, index, parent) {
       property.predicate == 'ForAllSuchThat' ||
       property.predicate == 'ThereExistsSuchThat'
     ) {
+      // quantifier
+      children[0] = searchInteractiveNode(
+        contracts,
+        property.inputDefs[0],
+        name,
+        index + 1,
+        newContract
+      )
+      // innerProperty
       children[2] = searchInteractiveNode(
         contracts,
         property.inputDefs[2],
@@ -64,7 +72,7 @@ function searchInteractiveNode(contracts, property, name, index, parent) {
         )
       })
     }
-    property.compiledChildren = children
+    newContract.statement.compiledChildren = children
     // If not atomic proposition, generate a contract
     contracts.push(newContract)
     return newContract
@@ -105,40 +113,6 @@ function getArguments(property) {
     })
   }
   return args
-}
-
-function isNotAtomicProposition(predicate) {
-  return (
-    predicate == 'ForAllSuchThat' ||
-    predicate == 'ThereExistsSuchThat' ||
-    predicate == 'Not' ||
-    predicate == 'And' ||
-    predicate == 'Or'
-  )
-}
-
-function genIsValidChallenge(property, name, index) {
-  if (property.predicate == 'ForAllSuchThat') {
-    property.isValidChallenge = {
-      predicate: 'Not',
-      inputs: [property.compiledChildren[0]],
-      challengeInput: property.inputDefs[1]
-    }
-  } else if (property.predicate == 'Not') {
-    property.isValidChallenge = property.compiledChildren[0]
-  } else if (property.predicate == 'And') {
-    property.isValidChallenge = {
-      predicate: 'Not',
-      inputs: property.compiledChildren,
-      challengeInput: 0
-    }
-  }
-  property.inputDefs.forEach(p => {
-    if (typeof p == 'object' && !!p.predicate) {
-      genIsValidChallenge(p, name, index + 1)
-    } else {
-    }
-  })
 }
 
 module.exports = {
